+++ src/gui/mapshaper-gui-export.js
@@ -6,13 +6,31 @@ gui.exportIsSupported = function() {
     !!window.navigator.msSaveBlob;
 };
 
+function canSaveToServer() {
+  return typeof mapshaper.manifest.allow_saving != 'undefined' &&
+    !!mapshaper.manifest.allow_saving && typeof fetch == 'function';
+}
+
 // replaces function from mapshaper.js
 internal.writeFiles = function(files, opts, done) {
   var filename;
   if (!utils.isArray(files) || files.length === 0) {
     done("Nothing to export");
+  } else if (canSaveToServer() && !opts.save_to_download_folder) {
+    saveFilesToServer(files, opts, function(err) {
+      var msg;
+      if (err) {
+        msg = "<b>Direct save failed</b><br>Reason: " + err + ".";
+        msg += "<br>Downloading output as a fallback.";
+        gui.alert(msg);
+        // fall back to standard method if saving to server fails
+        internal.writeFiles(files, {save_to_download_folder: true}, done);
+      } else {
+        done();
+      }
+    });
   } else if (files.length == 1) {
-    saveBlob(files[0].filename, new Blob([files[0].content]), done);
+    saveBlobToDownloadFolder(files[0].filename, new Blob([files[0].content]), done);
   } else {
     filename = utils.getCommonFileBase(utils.pluck(files, 'filename')) || "output";
     saveZipFile(filename + ".zip", files, done);
@@ -42,7 +60,7 @@ function saveZipFile(zipfileName, files, done) {
   function nextFile() {
     if (toAdd.length === 0) {
       zipWriter.close(function(blob) {
-        saveBlob(zipfileName, blob, done);
+        saveBlobToDownloadFolder(zipfileName, blob, done);
       });
     } else {
       var obj = toAdd.pop(),
@@ -52,7 +70,41 @@ function saveZipFile(zipfileName, files, done) {
   }
 }
 
-function saveBlob(filename, blob, done) {
+function saveFilesToServer(exports, opts, done) {
+  var paths = internal.getOutputPaths(utils.pluck(exports, 'filename'), opts);
+  var data = utils.pluck(exports, 'content');
+  var i = -1;
+  next();
+  function next(err) {
+    i++;
+    if (err) return done(err);
+    if (i >= exports.length) {
+      gui.alert('<b>Saved</b><br>' + paths.join('<br>'));
+      return done();
+    }
+    saveBlobToServer(paths[i], new Blob([data[i]]), next);
+  }
+}
+
+function saveBlobToServer(path, blob, done) {
+  var q = '?file=' + encodeURIComponent(path);
+  var url = window.location.origin + '/save' + q;
+  fetch(url, {
+    method: 'POST',
+    credentials: 'include',
+    body: blob
+  }).then(function(resp) {
+    if (resp.status == 400) {
+      return resp.text();
+    }
+  }).then(function(err) {
+    done(err);
+  }).catch(function(resp) {
+    done('connection to server was lost');
+  });
+}
+
+function saveBlobToDownloadFolder(filename, blob, done) {
   var anchor, blobUrl;
   if (window.navigator.msSaveBlob) {
     window.navigator.msSaveBlob(blob, filename);